Upgrade Unicorn JS transpiler from Babel 5 to Babel 6. - Make changes to get Babel working again -- Get past decorator problems with `babel-plugin-transform-decorators-legacy` -- Get Babel5-style export defaults with `babel-plugin-add-module-exports` --- Electron .remote() Client modules have to use new Babel6-style `.default` import - Upgrade most packages - Move Logger config out to own file - Match header license boilerplate style to other pages - etc.
