+++ unicorn/js/main/ModelService.js
@@ -1,28 +1,25 @@
-/* -----------------------------------------------------------------------------
- * Copyright © 2015, Numenta, Inc. Unless you have purchased from
- * Numenta, Inc. a separate commercial license for this software code, the
- * following terms and conditions apply:
- *
- * This program is free software: you can redistribute it and/or modify it
- * under the terms of the GNU Affero Public License version 3 as published by
- * the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero Public License for
- * more details.
- *
- * You should have received a copy of the GNU Affero Public License along with
- * this program. If not, see http://www.gnu.org/licenses.
- *
- * http://numenta.org/licenses/
- * -------------------------------------------------------------------------- */
+// Copyright © 2015, Numenta, Inc.  Unless you have purchased from
+// Numenta, Inc. a separate commercial license for this software code, the
+// following terms and conditions apply:
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU Affero Public License version 3 as published by the Free
+// Software Foundation.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE. See the GNU Affero Public License for more details.
+//
+// You should have received a copy of the GNU Affero Public License along with
+// this program.  If not, see http://www.gnu.org/licenses.
+//
+// http://numenta.org/licenses/
 
 import childProcess from 'child_process';
 import EventEmitter from 'events';
 import path from 'path';
-import UserError from './UserError';
 import system from 'os';
+import UserError from './UserError';
 
 // @todo https://bitbucket.org/anthony_tuininga/cx_freeze/issues/161
 const MODEL_RUNNER_PATH = path.join(
@@ -64,8 +61,8 @@ export class ModelNotFoundError extends UserError {
  * to Unicorn Backend Model Runner python and NuPIC processes.
  */
 export class ModelService extends EventEmitter {
-  constructor() {
-    super();
+  constructor(...args) {
+    super(...args);
     this._models = new Map();
     this._maxConcurrency = this._calculateMaxConcurrency();
   }
