+++ unicorn/package.json
@@ -74,9 +74,9 @@
     "bunyan": "1.5.1",
     "csv-streamify": "3.0.3",
     "dygraphs": "1.1.1",
-    "fbjs": "0.5.0",
+    "fbjs": "0.5.1",
     "fluxible": "1.0.3",
-    "fluxible-addons-react": "0.2.0",
+    "fluxible-addons-react": "0.2.2",
     "is-electron-renderer": "2.0.0",
     "js-csp": "0.5.0",
     "json-schema-instantiator": "0.3.0",
@@ -85,7 +85,7 @@
     "jsonschema": "1.0.2",
     "level-queryengine": "3.0.1",
     "level-sublevel": "6.5.2",
-    "leveldown": "1.4.2",
+    "leveldown": "1.4.3",
     "levelup": "1.3.1",
     "material-ui": "0.13.4",
     "moment": "2.10.6",
@@ -104,7 +104,7 @@
     "babel-eslint": "5.0.0-beta6",
     "babel-loader": "6.2.0",
     "babel-plugin-add-module-exports": "0.1.1",
-    "babel-plugin-transform-decorators-legacy": "1.2.0",
+    "babel-plugin-transform-decorators-legacy": "1.3.1",
     "babel-plugin-typecheck": "3.5.1",
     "babel-polyfill": "6.3.14",
     "babel-preset-es2015": "6.3.13",
@@ -130,6 +130,6 @@
     "style-loader": "0.13.0",
     "url-loader": "0.5.7",
     "webpack": "1.12.9",
-    "webpack-stream": "3.0.1"
+    "webpack-stream": "3.1.0"
   }
 }
