+++ unicorn/tests/js/mocha.opts
@@ -1,3 +1,3 @@
---recursive
 --compilers js:babel-core/register
+--recursive
 --timeout 5000
