+++ unicorn/tests/js/mocha.pipeline.opts
@@ -1,4 +1,4 @@
---recursive
 --compilers js:babel-core/register
+--recursive
 --reporter xunit
 --timeout 5000
